resume-agent/pyproject.toml
[project]
name = "resume-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
